petty_cash_final/petty_cash_final/myapp/resources.py

from import_export import resources
from .models import *

class SignupResource(resources.ModelResource):
    class Meta:
        model = Signup

class UserResource(resources.ModelResource):
    class Meta:
        model = User

class logResource(resources.ModelResource):
    class Meta:
        model = log