petty_cash_final/petty_cash_final/requirements.txt
asgiref==3.6.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.6
django-import-export==3.0.2
et-xmlfile==1.1.0
MarkupPy==1.14
numpy==1.24.1
odfpy==1.4.1
opencv-python==4.6.0.66
openpyxl==3.1.0
Pillow==9.4.0
PyYAML==6.0
sqlparse==0.4.3
tablib==3.3.0
tzdata==2022.7
xlrd==2.0.1
xlwt==1.3.0
